// src/model/CreateVideoCaptionResponse.ts
/**
 * @w3stream/nodejs-client
 * W3STREAM Service
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType.js';
import CreateVideoCaptionData from './CreateVideoCaptionData.js';

export default class CreateVideoCaptionResponse {
  'data'?: CreateVideoCaptionData;
  'status'?: string;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'data',
      baseName: 'data',
      type: 'CreateVideoCaptionData',
      format: '',
    },
    {
      name: 'status',
      baseName: 'status',
      type: 'string',
      format: '',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return CreateVideoCaptionResponse.attributeTypeMap;
  }
}